+++ plugin/phppoll/lang/en/realtimeplugin_phppoll.php
@@ -28,7 +28,7 @@ defined('MOODLE_INTERNAL') || die();
 $string['checkinterval'] = 'Check interval, ms';
 $string['checkintervaldesc'] = 'Interval between the polling requests to the server and also sleep interval between checks for new events on the server during long polling, in milliseconds, can not be less than {$a} ms';
 $string['pluginname'] = 'PHP polling';
-$string['privacy:metadata'] = 'The PHP polling plugin only stores user information for a short period of time';
+$string['privacy:metadata'] = 'The PHP polling plugin does not permanently store any personal data. It keeps real-time events in the database for a few minutes so that they can be delivered to the users\' browsers, and then deletes them.';
 $string['requesttimeout'] = 'Maximum duration for polling requests';
 $string['requesttimeoutdesc'] = 'Maximum duration of a polling request if there are no new events, set to 0 to use short polling instead of long polling';
 $string['taskcleanup'] = 'Clean-up for events in PHP polling';